range: fix size hint
`size_hint` was implicitly assuming that for `t: T` the following holds:

    t + One::one() != t

This is not true, i.e. if `T` is a float.
Due to this fact, `range` can return an infinite iterator. However,
`size_hint` would always claim the iterator is finite.

This commit fixes `size_hint` by not specifying an upper bound.

Updates #17010.